requirements.txt
streamlit
langchain
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
python-dotenv
pypdf
docx2txt
unstructured
rank-bm25
ragas
datasets
openai
fpdf
tiktoken